==> 1_web_scraping_1.ipynb <==
# Kod ten wchodzi na podaną stronę, pobiera wyznaczone linki i przechodzi do kolejnej strony, z której także pobiera linki i tak dalej.
# Przejście do kolejnej strony jest tu utrudnione, ponieważ nie ma oznaczonego konkretnego adresu URL.

# Importy
import requests
from bs4 import BeautifulSoup
import pandas as pd

initial_url = "https://lubimyczytac.pl/oficjalne-recenzje-ksiazek?page=1&listId=listOfArticles&categoryCms[]=2&time[]=2014-01-01&time[]=2023-12-31&timeFilter=4&timeFilterRange=01-2014,12-2023&sortBy=youngest&paginatorType=Standard"

def get_review_links(initial_url):
    review_links = []

    # Tworzę początek adresu kolejnej strony
    base_url = 'https://lubimyczytac.pl/oficjalne-recenzje-ksiazek?page='
    #Tworzę koniec adresu kolejnej strony
    query_params = '&listId=listOfArticles&categoryCms[]=2&time[]=2014-01-01&time[]=2023-12-31&timeFilter=4&timeFilterRange=01-2014,12-2023&sortBy=youngest&paginatorType=Standard'

    # Tworzę listę zawierającą numery stron, z których będe pobierała linki
    lista_pages = list(range(1, 400))

    for page_num in lista_pages:
        #Aktualizuję URL do bieżącej strony
        current_url = base_url + str(page_num) + query_params

        # Pobieram zawartość strony
        response = requests.get(current_url)
        soup = BeautifulSoup(response.text, 'html.parser')

        # Szukam linki do recenzji na bieżącej stronie
        review_links_on_page = soup.find_all('a', class_='authorAllBooks__singleTextTitle')
        review_links.extend(link['href'] for link in review_links_on_page)

    return review_links

# Pobieram linki do recenzji
review_links = get_review_links(initial_url)

# Tworzę obiekt DataFrame z listy linków
df = pd.DataFrame({'Link do recenzji': review_links})

# Zapisuję DataFrame do pliku Excela
file_name = 'review_links_books.xlsx'
df.to_excel(file_name, index=False)
