==> 2_web_scraping_2.ipynb <==

# Jeśli linki pobrane w poprzednim kroku są tylko fragmentem całego linku, to tworzymy całe linki

import openpyxl
import pandas as pd

# Wczytuję pierwszy plik Excel
df = pd.read_excel("review_links_books.xlsx")


# Tworzę zmienną z początkiem linku
base_url = "https://lubimyczytac.pl"

# Dodaję początek linku do fragmentu link
link = base_url + df["Link do recenzji"]


# Zapisuję zmieniony plik xlsx
link.to_excel("review_links_books_1a.xlsx", index=False)

print("Done")
