==> 3_web_scraping_3.ipynb <==
# Kod ten otwiera wskazany plik .xlsx.
# Pobiera link z pierwszego wiersza, otwiera go, pobiera wskazane dane i zapisuje je do nowego pliku .csv.
# I przechodzi do kolejnego linku z pierwszego pliku .xlsx.

# Importy bibliotek
import requests
from bs4 import BeautifulSoup
import pandas as pd
from time import sleep
from datetime import datetime

# Otwieram plik z zapisanymi linkami do recenzji
excel_file = "review_links_books_1a.xlsx"
# Tworzę ramkę danych
df = pd.read_excel(excel_file)

reviews = []

for link in df["Link do recenzji"]:
    print(link)
    # Pobieram stronę z recenzją
    headers = {"User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_10_1) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/39.0.2171.95 Safari/537.36"}
    response = requests.get(link, headers=headers)
    sleep(3)
    content = response.text
    soup = BeautifulSoup(content, "html.parser")

    # Pobieram autora recenzowanej książki
    try:
        author = str(soup.find("div", {"class": "authorAllBooks__singleTextAuthor"}).text.strip())
    except (AttributeError, ValueError):
        author = ''

    # Pobieram tytuł recenzowanej książki
    try:
        title = str(soup.find("a", {"class": "authorAllBooks__singleTextTitle float-left"}).text.strip())  ##można też zapisać tak: soup.find("a", class_="authorAllBooks__singleTextTitle float-left").text.strip()
    except (AttributeError, ValueError):
        title = ""

    # Pobieram datę recenzji
    try:

        date_text = soup.find("span", {"class": "newsCats__date grey d-flex align-items-center"}).text.strip()
        # Przygotowuję pobrany tekst do przeformatowania
        # strptime: Jest to metoda klasy datetime, która służy do przekształcania tekstu w formie daty w obiekt daty i czasu.
        # Skrót strptime oznacza "string parse time".
        date_parts = datetime.strptime(date_text, "%d.%m.%Y")
        # Przekształcam datę na nowy format
        date = date_parts.strftime("%Y-%m-%d")
    except (AttributeError, ValueError):
        date = ""

    # Pobieram treść recenzji
    try:
        text1 = str(soup.find("div", {"class": "articlesCmsContent__header"}).text.strip())
    except (AttributeError, ValueError):
        text1 = ""
    try:
        text2 = str(soup.find("div", {"class": "articlesCmsContent__text--narrow col-lg-8 offset-lg-2 px-sm-0 px-3 mx-sm-auto mb-3"}).text.strip())
    except (AttributeError, ValueError):
        text2 = ""

    text = text1 + "" "" + text2

    # Pobieram średnią ocen
    try:
        rating_text = soup.find("span", {"class": "listLibrary__ratingStarsNumber"}).text.strip()
        rating_float = float(rating_text.replace(',', '.'))
        rating  = rating_float
    except (AttributeError, ValueError):
        rating = ""

    # Pobieram liczbę ocen
    try:
        rating_all_text = soup.find("div", {"class": "listLibrary__ratingAll"}).text.strip() #np. '74 ocen'
        #zawartość elementu div dzielę po spacji
        rating_all_parts = rating_all_text.split(" ") #'74', 'ocen'
        try:
            rating_all = int(rating_all_parts[0]) #tylko pierwszy fragment z numerem
        except (AttributeError, ValueError):
            rating_all = ""
    except (AttributeError, ValueError):
        rating_all = ""

    # Pobieram liczbę czytelników
    try:
        readers_all_text = soup.find("span", {"class": "small grey mr-2 mb-3"}).text.strip() #np. 'Czytelnicy: 257'
        #zawartość elementu span dzielę po spacji
        readers_all_parts = readers_all_text.split(" ") # 'Czytelnicy:', '257'
        try:
            readers_all = int(readers_all_parts[1]) #tylko drugi fragment z numerem
        except (AttributeError, ValueError):
            readers_all = ""
    except (AttributeError, ValueError):
        readers_all = ""

    reviews.append({"Link": link,
                      "Autor" : author,
                      "Tytuł" : title,
                      "Data_recenzji" : date,
                      "Treść_recenzji" : text,
                      "Średnia_ocen" : rating,
                      "Liczba_ocen" : rating_all,
                      "Liczba_czytelników" : readers_all})

df_reviews = pd.DataFrame(reviews)
df_reviews.to_csv("reviews_all.csv")

df_reviews = pd.DataFrame(reviews)
df_reviews.to_csv("reviews_all.csv")
