==> 6_modelowanie_etykiety_huggingface.ipynb <==
# ANALIZA SENTYMENTU - 4 etykiety

# Importy
import pandas as pd
from sklearn.model_selection import train_test_split
from transformers import AutoTokenizer, AutoModelForSequenceClassification, AdamW, DataCollatorWithPadding, TrainingArguments, Trainer
from torch.utils.data import DataLoader, TensorDataset
import torch
from sklearn.metrics import accuracy_score, classification_report
from datasets import Dataset
import datasets
import accelerate
from imblearn.over_sampling import SMOTE

df = pd.read_csv("reviews_eda1b.csv")

df.info()

df.columns

# Pierwszy model, który przetestuje to będzie RandomForestClassifier
# W związku z tym, że mam wyraźny problem niezrównoważonych klas, będe musiała zastosować technike oversamplingu.
# Nie wiem czy duplikowanie danych to będzie najlepsze rozwiązanie. 
# Najpierw więc zastosuję SMOTE (Synthetic Minority Over-sampling Technique)

from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import classification_report
from imblearn.over_sampling import SMOTE

# Przygotowuję zbiory: treningowy, walidacyjny i testowy
train_data, test_data = train_test_split(df, test_size=0.2, random_state=42)
train_data, val_data = train_test_split(train_data, test_size=0.25, random_state=42)

# Tworzę macierz TF-IDF
vectorizer = TfidfVectorizer(max_features=5000, ngram_range=(1,2))
X_train = vectorizer.fit_transform(train_data["tokenized_reviews"])
X_val = vectorizer.transform(val_data["tokenized_reviews"])
X_test = vectorizer.transform(test_data["tokenized_reviews"])

# Etykiety
y_train = train_data["sentyment"]
y_val = val_data["sentyment"]
y_test = test_data["sentyment"]

# Podejście 1 - model RandomForestClassifier
rf_model_base = RandomForestClassifier(random_state=42)
rf_model_base.fit(X_train, y_train)

# Teraz będę oceniać modelu bazowego na zestawie walidacyjnym
y_val_pred_base = rf_model_base.predict(X_val)
print("Classification Report dla modelu bazowego:\n", classification_report(y_val, y_val_pred_base))


# W związku z dość kiepskimi wynikami, teraz zastosuję SMOTE

# SMOTE do zestawu treningowego
smote = SMOTE(random_state=42)
X_train_resampled, y_train_resampled = smote.fit_resample(X_train, y_train)

# Model z zastosowaniem SMOTE
rf_model_resampled = RandomForestClassifier(random_state=42)
rf_model_resampled.fit(X_train_resampled, y_train_resampled)


# Ocena modelu z zastosowaniem SMOTE na zestawie walidacyjnym
y_val_pred_resampled = rf_model_resampled.predict(X_val)
print("Classification Report dla modelu z zastosowaniem SMOTE:\n", classification_report(y_val, y_val_pred_resampled))

# Model XGBoost

# base

import xgboost as xgb
from sklearn.metrics import accuracy_score

# Utworzenie i trening modelu XGBoost
xgb_model_base = xgb.XGBClassifier()
xgb_model_base.fit(X_train, y_train)

# Ocena modelu na danych walidacyjnych
y_pred_xgb = xgb_model_base.predict(X_test)
print("Classification Report XGBoost:")
print(classification_report(y_test, y_pred_xgb))

# SMOTE

# Utworzenie i trening modelu XGBoost na danych po SMOTE
xgb_model_smote = xgb.XGBClassifier()
xgb_model_smote.fit(X_train_resampled, y_train_resampled)

# Ocena modelu XGBoost po SMOTE na danych testowych
y_pred_xgb_smote = xgb_model_smote.predict(X_val)
print("Classification Report XGBoost with SMOTE:")
print(classification_report(y_test, y_pred_xgb_smote))

# Teraz LSTM

from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Embedding, LSTM, Dense, SpatialDropout1D, GlobalAveragePooling1D
from tensorflow.keras.callbacks import EarlyStopping
import tensorflow as tf
from sklearn.feature_extraction.text import TfidfVectorizer

# Dzielę dane na zbiory train, val, test
train_data, test_data = train_test_split(df, test_size=0.2, random_state=42)
train_data, val_data = train_test_split(train_data, test_size=0.25, random_state=42)


# Tworzę macierz TF-IDF
vectorizer = TfidfVectorizer(max_features=5000, ngram_range=(1, 2))
X_train = vectorizer.fit_transform(train_data["tokenized_reviews"])
X_val = vectorizer.transform(val_data["tokenized_reviews"])
X_test = vectorizer.transform(test_data["tokenized_reviews"])

# Przekształcam macierze rzadkie na gęste przed porządkowaniem indeksów tf.sparse.reorder
X_train_dense = X_train.toarray()
X_val_dense = X_val.toarray()
X_test_dense = X_test.toarray()

# Tworzę SparseTensor z gęstych macierzy
X_train_sparse = tf.sparse.reorder(tf.sparse.from_dense(X_train_dense))
X_val_sparse = tf.sparse.reorder(tf.sparse.from_dense(X_val_dense))
X_test_sparse = tf.sparse.reorder(tf.sparse.from_dense(X_test_dense))

# Etykiety
y_train = train_data["sentyment"]
y_val = val_data["sentyment"]
y_test = test_data["sentyment"]

# Definicja parametrów
max_words = 5000  # maksymalna liczba unikalnych słów
embedding_dim = 128  # wymiar przestrzeni osadzeń


# Budowa modelu
model = Sequential()
model.add(Embedding(input_dim=max_words, output_dim=embedding_dim, input_length=None))
model.add(SpatialDropout1D(0.2))
model.add(LSTM(100, dropout=0.2, recurrent_dropout=0.2, return_sequences=True))
model.add(GlobalAveragePooling1D())  # GlobalAveragePooling1D zamiast Flatten
model.add(Dense(4, activation='softmax'))  # 4 klasy sentymentu

# Kompilacja modelu
model.compile(loss='sparse_categorical_crossentropy', optimizer='adam', metrics=['accuracy'])


# Wczesne zatrzymywanie
early_stopping = EarlyStopping(monitor='val_loss', patience=3, restore_best_weights=True)


# Trenowanie modelu
history = model.fit(X_train_sparse, y_train, epochs=10, batch_size=64, validation_data=(X_val_sparse, y_val), callbacks=[early_stopping])


# Ocena modelu
accuracy = model.evaluate(X_test_sparse, y_test)[1]
print(f'Test Accuracy: {accuracy}')
