==> 6a_modelowanie_etykiety_huggingface.ipynb <==
# Model BertForSequenceClassification na danych z 4 etykietami

# Importy
import pandas as pd
from transformers import BertTokenizer, BertForSequenceClassification, AdamW
from torch.utils.data import DataLoader, TensorDataset
from sklearn.model_selection import train_test_split
import torch
from sklearn.metrics import classification_report, accuracy_score
from imblearn.over_sampling import SMOTE


df = pd.read_csv("reviews_eda1b.csv")

df.columns

# Moje dane
texts = df["Treść_recenzji"].tolist()  # Lista tekstów
labels = df["sentyment"].tolist()  # Lista etykiet sentymentu

texts

base

# Dzielę na zbiór treningowy, walidacyjny i testowy
train_texts, test_texts, train_labels, test_labels = train_test_split(texts, labels, test_size=0.2, random_state=42)
train_texts, val_texts, train_labels, val_labels = train_test_split(train_texts, train_labels, test_size=0.1, random_state=42)

# Tokenizacja tekstu
tokenizer = BertTokenizer.from_pretrained('google-bert/bert-base-multilingual-cased')

tokenized_train = tokenizer(train_texts, padding=True, truncation=True, return_tensors="pt")
tokenized_val = tokenizer(val_texts, padding=True, truncation=True, return_tensors="pt")
tokenized_test = tokenizer(test_texts, padding=True, truncation=True, return_tensors="pt")

# Konwersja etykiet do tensorów
train_labels = torch.tensor(train_labels)
val_labels = torch.tensor(val_labels)
test_labels = torch.tensor(test_labels)

# Muszę zrobić TensorDataset
train_dataset = TensorDataset(tokenized_train.input_ids, tokenized_train.attention_mask, tokenized_train.token_type_ids, train_labels)
val_dataset = TensorDataset(tokenized_val.input_ids, tokenized_val.attention_mask, tokenized_val.token_type_ids, val_labels)
test_dataset = TensorDataset(tokenized_test.input_ids, tokenized_test.attention_mask, tokenized_test.token_type_ids, test_labels)

#jaikes błedy, więc robię  DataLoader
train_dataloader = DataLoader(train_dataset, batch_size=8, shuffle=True)
val_dataloader = DataLoader(val_dataset, batch_size=8, shuffle=False)
test_dataloader = DataLoader(test_dataset, batch_size=8, shuffle=False)

# Inicjalizuję model
model = BertForSequenceClassification.from_pretrained('google-bert/bert-base-multilingual-cased', num_labels=4)
optimizer = AdamW(model.parameters(), lr=5e-5)

# Trenowanie modelu
num_epochs = 3

for epoch in range(num_epochs):
    print(epoch)
    model.train()
    for batch in train_dataloader:
        inputs = {'input_ids': batch[0], 'attention_mask': batch[1], 'token_type_ids': batch[2], 'labels': batch[3]}
        outputs = model(**inputs)
        loss = outputs.loss
        loss.backward()
        optimizer.step()
        optimizer.zero_grad()

# Ewaluacja modelu
model.eval()
all_predictions = []
all_true_labels = []

with torch.no_grad():
    for batch in test_dataloader:
        inputs = {'input_ids': batch[0], 'attention_mask': batch[1], 'token_type_ids': batch[2]}
        outputs = model(**inputs)
        predictions = torch.argmax(outputs.logits, dim=1)
        all_predictions.extend(predictions.tolist())
        all_true_labels.extend(batch[3].tolist())


# wyniki
print(classification_report(all_true_labels, all_predictions))
print("Accuracy:", accuracy_score(all_true_labels, all_predictions))

smote

from transformers import BertTokenizer, BertForSequenceClassification, AdamW
from torch.utils.data import DataLoader, TensorDataset
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import LabelEncoder
from imblearn.over_sampling import SMOTE
import torch
from sklearn.metrics import classification_report, accuracy_score

# Dziele na zbiór treningowy, walidacyjny i testowy
train_texts, test_texts, train_labels, test_labels = train_test_split(texts, labels, test_size=0.2, random_state=42)
train_texts, val_texts, train_labels, val_labels = train_test_split(train_texts, train_labels, test_size=0.1, random_state=42)

# Tokenizacja tekstu
tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')

tokenized_train = tokenizer(train_texts, padding=True, truncation=True, return_tensors="pt")
tokenized_val = tokenizer(val_texts, padding=True, truncation=True, return_tensors="pt")
tokenized_test = tokenizer(test_texts, padding=True, truncation=True, return_tensors="pt")


# Oversampling SMOTE
# były błedy, więc muszę przekształcic dane treningowe do postaci 2D array
X_train_smote = torch.cat([tokenized_train.input_ids, tokenized_train.attention_mask, tokenized_train.token_type_ids], dim=1).numpy()
smote = SMOTE(random_state=42, sampling_strategy='auto')
X_train_resampled, y_train_resampled = smote.fit_resample(X_train_smote, train_labels.numpy())


# Dzielę dane z powrotem na tokeny i etykiety
tokenized_train_resampled = torch.tensor(X_train_resampled[:, :tokenized_train.input_ids.shape[1]]), \
                            torch.tensor(X_train_resampled[:, tokenized_train.input_ids.shape[1]:2*tokenized_train.input_ids.shape[1]]), \
                            torch.tensor(X_train_resampled[:, 2*tokenized_train.input_ids.shape[1]:]), \
                            torch.tensor(y_train_resampled)


# Tworzę DataLoader dla danych po oversamplingu
train_dataloader_resampled = DataLoader(TensorDataset(*tokenized_train_resampled), batch_size=8, shuffle=True)

# Oversampling SMOTE dla danych walidacyjnych 
X_val_smote = torch.cat([tokenized_val.input_ids, tokenized_val.attention_mask, tokenized_val.token_type_ids], dim=1).numpy()
X_val_resampled, y_val_resampled = smote.fit_resample(X_val_smote, val_labels.numpy())

tokenized_val_resampled = torch.tensor(X_val_resampled[:, :tokenized_val.input_ids.shape[1]]), \
                          torch.tensor(X_val_resampled[:, tokenized_val.input_ids.shape[1]:2*tokenized_val.input_ids.shape[1]]), \
                          torch.tensor(X_val_resampled[:, 2*tokenized_val.input_ids.shape[1]:]), \
                          torch.tensor(y_val_resampled)

val_dataloader_resampled = DataLoader(TensorDataset(*tokenized_val_resampled), batch_size=8, shuffle=False)

# Oversampling SMOTE dla danych testowych 
X_test_smote = torch.cat([tokenized_test.input_ids, tokenized_test.attention_mask, tokenized_test.token_type_ids], dim=1).numpy()
X_test_resampled, y_test_resampled = smote.fit_resample(X_test_smote, test_labels.numpy())

tokenized_test_resampled = torch.tensor(X_test_resampled[:, :tokenized_test.input_ids.shape[1]]), \
                           torch.tensor(X_test_resampled[:, tokenized_test.input_ids.shape[1]:2*tokenized_test.input_ids.shape[1]]), \
                           torch.tensor(X_test_resampled[:, 2*tokenized_test.input_ids.shape[1]:]), \
                           torch.tensor(y_test_resampled)

test_dataloader_resampled = DataLoader(TensorDataset(*tokenized_test_resampled), batch_size=8, shuffle=False)

# Konwersja etykiet do tensorów
train_labels = torch.tensor(train_labels)
val_labels = torch.tensor(val_labels)
test_labels = torch.tensor(test_labels)

from torch.optim import Adam

# Inicjalizuję modelu
model = BertForSequenceClassification.from_pretrained('bert-base-uncased', num_labels=4)
optimizer = Adam(model.parameters(), lr=5e-5)

# Trening modelu
num_epochs = 3

for epoch in range(num_epochs):
    print(epoch)
    model.train()
    for batch in train_dataloader_resampled:
        inputs = {'input_ids': batch[0], 'attention_mask': batch[1], 'token_type_ids': batch[2], 'labels': batch[3]}
        outputs = model(**inputs)
        loss = outputs.loss
        loss.backward()
        optimizer.step()
        optimizer.zero_grad()

# Ewaluacja modelu
model.eval()
all_predictions = []
all_true_labels = []

with torch.no_grad():
    for batch in val_dataloader_resampled:
        inputs = {'input_ids': batch[0], 'attention_mask': batch[1], 'token_type_ids': batch[2]}
        outputs = model(**inputs)
        predictions = torch.argmax(outputs.logits, dim=1)
        all_predictions.extend(predictions.tolist())
        all_true_labels.extend(batch[3].tolist())

# Wyniki
print(classification_report(all_true_labels, all_predictions))
print("Accuracy:", accuracy_score(all_true_labels, all_predictions))
